requirements.txt
Django==3.0.8
PyMySQL==0.10.0
PySocks==1.7.1
asgiref==3.2.10
attrs==20.1.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
country-currencies==0.2
currency-exchange==1.0.0
decorator==4.4.2
django-countries==6.1.3
django-crispy-forms==1.9.2
exchange==0.3
geographiclib==1.50
geojson==2.5.0
geopy==2.0.0
idna==2.10
importlib-metadata==1.7.0
ipython-genutils==0.2.0
jsonschema==3.2.0
jupyter-core==4.6.3
mysql-client==0.0.1
nbformat==5.0.7
oweatherm==0.5.2
pigar==0.10.0
pip==20.2.2
pyowm==3.0.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-gettext==4.0
pytz==2020.1
pywin32==228
requests==2.24.0
setuptools==49.6.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
weather-forecast==0.49
zipp==3.1.0